[Increment] - Finish level 7 (#2)
* Enhancement : Change bufferedReader to scanner, move it inside CommandProcessorImpl.java

* Enhancement : Make CommandProcessorImpl Autoclosable, so the scanner would be closed in case of exception

* Increment: Finish level-7 for persistence data storage using a local
file

* Increment: Finish level-7 for persistence data storage using a local
file
