--- src/main/java/commandInterface/ICommandProcessor.java
@@ -1,13 +1,0 @@
-package commandInterface;
-
-/**
- * Command line processor interface that takes in the input from the console
- */
-public interface ICommandProcessor {
-	/**
-	 * Parse the input from console into corresponding command and process it, current implementation using Command Processor
-	 *
-	 * @param input String containing the lines
-	 */
-	void processInput(String input);
-}
